Fixed bug that prevented signing up for a ride because their number was greater than the maximum Int value, replaced deprecated String selectors with new Swift ones.

+++ Cru/FakeServerClient.swift
@@ -177,7 +177,7 @@ class FakeServerClient: ServerProtocol {
         handler(true)
     }
     
-    func checkIfValidNum(num: Int, handler: (Bool)->()){
+    func checkIfValidNum(num: Double, handler: (Bool)->()){
         
     }
