+++ Cru/NewOfferRideViewController.swift
@@ -522,7 +522,7 @@ class NewOfferRideViewController: UIViewController, UITextFieldDelegate, UIPopov
     //Send ride request
     func sendRideOffer(){
         if parsedNum != nil {
-            CruClients.getServerClient().checkIfValidNum(Int(parsedNum!)!, handler: postRideOffer)
+            CruClients.getServerClient().checkIfValidNum(Double(parsedNum!)!, handler: postRideOffer)
         }
         
     }
